Working on rider issues
#926

+++ Src/CSharpier.Rider/src/main/java/com/intellij/csharpier/CustomPathInstaller.java
@@ -7,48 +7,68 @@ import java.io.File;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.regex.Pattern;
 
 public class CustomPathInstaller {
     Logger logger = CSharpierLogger.getInstance();
+    String customPath;
 
-    public void ensureVersionInstalled(String version) throws Exception {
+    public CustomPathInstaller(CSharpierSettings settings) {
+        this.customPath = settings.getCustomPath();
+    }
+
+    public boolean ensureVersionInstalled(String version) throws Exception {
         if (version == null || version.equals("")) {
-            return;
+            return true;
         }
+        if (this.customPath != "" && this.customPath != null) {
+            this.logger.debug("Using csharpier at a custom path of " + this.customPath);
+            return true;
+        }
+
         var pathToDirectoryForVersion = getDirectoryForVersion(version);
         var directoryForVersion = new File(pathToDirectoryForVersion);
         if (directoryForVersion.exists()) {
-            try {
-                Map<String, String> env = new HashMap<>();
-                env.put("DOTNET_NOLOGO", "1");
-
-                var command = new String[] { getPathForVersion(version), "--version" };
-                var output = ProcessHelper.ExecuteCommand(command, env, new File(pathToDirectoryForVersion));
-
-                this.logger.debug("dotnet csharpier --version output: " + output);
-
-                if (output.equals(version))
-                {
-                    this.logger.debug("CSharpier at " + pathToDirectoryForVersion + " already exists");
-                    return;
-                }
-            }
-            catch (Exception ex) {
-                // TODO somehow I got a bunch of the versions to install that were missing dotnet-csharpier in the root of the custom path
-                // this needs to do a better job of figuring that out and reporting it.
-                // I think when this fails to install, then the other stuff gets stuck in an infinite loop
-                logger.warn("Exception while running 'dotnet csharpier --version' in " + pathToDirectoryForVersion, ex);
+            if (this.validateInstall(pathToDirectoryForVersion, version)) {
+                return true;
             }
 
-            // if we got here something isn't right in the current directory
+            this.logger.debug(
+                    "Removing directory at " + pathToDirectoryForVersion + " because it appears to be corrupted"
+            );
             deleteDirectory(directoryForVersion);
         }
 
         var command = new String[]{"dotnet", "tool", "install", "csharpier", "--version", version, "--tool-path", pathToDirectoryForVersion};
         ProcessHelper.ExecuteCommand(command, null, null);
+
+        return this.validateInstall(pathToDirectoryForVersion, version);
     }
 
-    boolean deleteDirectory(File directoryToBeDeleted) {
+    private boolean validateInstall(String pathToDirectoryForVersion, String version) {
+        try {
+            Map<String, String> env = new HashMap<>();
+            env.put("DOTNET_NOLOGO", "1");
+
+            var command = new String[] { this.getPathForVersion(version), "--version" };
+            var output = ProcessHelper.ExecuteCommand(command, env, new File(pathToDirectoryForVersion)).trim();
+
+            this.logger.debug("dotnet csharpier --version output: " + output);
+
+            if (output.split(Pattern.quote("+"))[0].equals(version))
+            {
+                this.logger.debug("CSharpier at " + pathToDirectoryForVersion + " already exists");
+                return true;
+            }
+        }
+        catch (Exception ex) {
+            this.logger.warn("Exception while running 'dotnet csharpier --version' in " + pathToDirectoryForVersion, ex);
+        }
+
+        return false;
+    }
+
+    private boolean deleteDirectory(File directoryToBeDeleted) {
         File[] allContents = directoryToBeDeleted.listFiles();
         if (allContents != null) {
             for (File file : allContents) {
@@ -59,6 +79,10 @@ public class CustomPathInstaller {
     }
 
     private String getDirectoryForVersion(String version) throws Exception {
+        if (this.customPath != "" && this.customPath != null) {
+            return this.customPath;
+        }
+
         if (SystemUtils.IS_OS_WINDOWS) {
             return Path.of(System.getenv("LOCALAPPDATA"), "CSharpier", version).toString();
         }
